Update deploy scripts. bcrypt -> bcryptjs.
Azure App Service doesn't support bcrypt out of the box due to running node in 32-bit. To change this it would be necessary to add a 64-bit node.exe to the instance and a web.config which specifies the node executable to use. For now it seems simpler to go with bcryptjs and avoid the native module dependency.